move pySite class to its corresponding cpp and h files
